including joptimizer

+++ at.fhv.smartdevices/src/at/fhv/smartdevices/helper/MapHelper.java
@@ -5,25 +5,25 @@ import java.util.SortedMap;
 import at.fhv.smartdevices.commons.SerializableTreeMap;
 
 public class MapHelper {
-	public static <K extends Number, V>  float[] keysToDoubleArray(SortedMap<K, V> map)
+	public static <K extends Number, V>  double[] keysToDoubleArray(SortedMap<K, V> map)
 	{
 		int amount = map.keySet().size();
 		Number[] keys = (Number[]) map.keySet().toArray(new Number[amount]);	
-		float[] retVal = new float[amount];
+		double[] retVal = new double[amount];
 		for (int i=0;i<amount;i++) {
-			retVal[i] = keys[i].floatValue();
+			retVal[i] = keys[i].doubleValue();
 		}		
 		return retVal;		
 	}		
 	
-	public static <K, V extends Number>  float[] valuesToDoubleArray(SortedMap<K, V> map)
+	public static <K, V extends Number>  double[] valuesToDoubleArray(SortedMap<K, V> map)
 	{
 		int amount = map.values().size();
 		Number[] values = (Number[]) map.values().toArray(new Number[amount]);
 		
-		float[] retVal = new float[amount];
+		double[] retVal = new double[amount];
 		for (int i=0;i<amount;i++) {			
-			retVal[i] = values[i].floatValue();
+			retVal[i] = values[i].doubleValue();
 		}		
 		return retVal;		
 	}
